src/fabrik-proxy/requirements.txt
Flask==2.3.3
pika==1.3.2
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-pika==0.41b0
opentelemetry-exporter-otlp-proto-http==1.20.0
requests==2.31.0
